% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acronymr.R
\name{ac}
\alias{ac}
\title{Acronym handling function}
\usage{
ac(acronym = NULL, acronyms = "acronyms", p = FALSE,
  altFrm = FALSE, lngFrm = FALSE)
}
\arguments{
\item{acronym}{character string matching value in Acronym column of \code{acronyms}}

\item{acronyms}{character string containing name of acronym data frame}

\item{p}{logical indicating if the acronym is plural}

\item{altFrm}{logical indicating if alternate form of acronym should be used}

\item{lngFrm}{logical indicating if long form of acronym should be used}
}
\value{
When an acronym is used, the \code{acronym} entry is found in \code{acronmys} and if Include is set to \code{FALSE} indicating first usage, the return value is the long form of the acronym followed by either the acronym name or alternate form depending on \code{altFrm} enclosed in parentheses. \code{acronyms$Include} is set to \code{TRUE}. Subsequent usage of an acronym will return the acronym name or alternate form.

If \code{acronyms} is not found or if \code{acronyms} exists but is not a data frame or does not include the necessary columns, the function stops and issues an error. If the acronym is not found in \code{acronyms} a warning is issued and XXXacro:Acronym Not DefinedXXX is returned.
}
\description{
\code{ac} formats acronyms when called inline in rmarkdown.
}
\details{
\code{ac} handles inline acronyms and references a dataframe with the acronym name, alternate formatting, definition and include flag and returns a character string. Options include whether the acronym usage is singular or plural and whether the full definition should be used. The function handles missing acronyms, checks for first time usage, and formats the return accordingly. This function is most commonly used in rmarkdown documents.

The \code{acronyms} data frame must contain the following columns:
\itemize{
  \item{Acronym = (character) name of acronym}
  \item{Alternate = (character) alternate formatting for use when the acronym uses special characters or other formatting such as super/subscripts, italics, etc.}
  \item{Definition = (character) long form of acronym}
  \item{Include = (logical) flag indicating if the acronym has been used in the document and must be reset to FALSE at the beginning of each render}
}
}
\examples{
# Example acronyms data frame
acronyms <-
  tibble::tribble(
    ~Acronym, ~Alternate, ~Definition,
    "IMO", NA,"In My Opinion",
    "PCR", NA, "Polymerase Chain Reaction",
    "RNA", NA, "Ribonucleic Acid",
    "H2NO3", "H~2~NO~3~", "Nitric Acid")

# Add inclusion flag
acronyms$Include <- FALSE
# Alphabetize acronyms for ease of reading
acronyms <- acronyms[order(acronyms$Acronym),]

# inline usage
`r ac(acronym = "IMO", acronyms = "acronyms", p = FALSE, altFrm = FALSE, lngFrm = FALSE)`
# or simply
` ac("IMO")`
}
