% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acronymr.R
\name{acTable}
\alias{acTable}
\title{Return acronyms used dataframe}
\usage{
acTable(acronyms = "acronyms", altFrm = TRUE, Caption = NULL,
  Tag = FALSE)
}
\arguments{
\item{acronyms}{name of acronym data frame as character string}

\item{altFrm}{logical indicating whether to use alternate form where available}

\item{Caption}{acronmy table caption}
}
\value{
If acronyms are used in the document, a dataframe. If \code{acronyms} does not exist or no acronyms were used in the document, a message is raised and returns \code{NULL}.
}
\description{
Create an acronym table of acronyms used and insert into text.
}
\details{
\code{acTable} is most often used within a knitted document. The function checks if an \code{acronyms} exists and whether any acronyms were used in the document. If both criteria are true, a dataframe containing the acronym and definition of all acronyms used in the document is returned.
}
\examples{
# This example assumes use within an rmarkdown documnet
acronyms <-
  tibble::tribble(
    ~Acronym, ~Alternate, ~Definition,
    "IMO", NA,"In My Opinion",
    "PCR", NA, "Polymerase Chain Reaction",
    "RNA", NA, "Ribonucleic Acid",
    "H2NO3", "H~2~NO~3~", "Nitric Acid")

# Add inclusion flag
acronyms$Include <- FALSE
# Alphabetize acronyms for ease of reading
acronyms <- acronyms[order(acronyms$Acronym),]

# place an acronym inline
`r ac("IMO")`

# generate acronym table
acTable("acronyms")
}
